requirements.txt
streamlit==1.26.0
yfinance==0.2.40
plotly.express==0.4.1
streamlit_extras==0.4.2